calculator-backend/requirements.txt
fastapi==0.133.0
uvicorn[standard]==0.41.0
python-dateutil==2.9.0.post0
